# be/swagger.json
{
    "info": {
        "title": "Taxi Booking API",
        "version": "1.0.0",
        "description": "RESTful API description with Swagger"
    },
    "host": "localhost:3000",
    "swagger": "2.0",
    "paths": {
...
        "/ride/accept-ride-request": {
            "post": {
                "tags": [
                    "Ride"
                ],
                "summary": "Driver can accept the ride request by passing the ride_id, Ride request appears on his/her(driver) device.",
                "description": null,
                "operationId": "acceptRideRequest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "value":"Bearer "
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/acceptRideRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Accept ride request failed"
                    },
                    "404": {
                        "description": "Accept ride request failed"
                    }
                }
            }
        },
...
        "/ride/ride-started-by-driver": {
            "post": {
                "tags": [
                    "Ride"
                ],
                "summary": "When a driver start ride then this api update database accordingly",
                "description": null,
                "operationId": "rideStartedByDriver",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "value":"Bearer "
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rideStartedByDriver"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Ride start request failed"
                    },
                    "404": {
                        "description": "Ride start request failed"
                    }
                }
            }
        },
        "/ride/update-ride-root": {
            "post": {
                "tags": [
                    "Ride"
                ],
                "summary": "Ride started and after each specific time interval this API will update the root lat long value",
                "description": null,
                "operationId": "updateRideRoot",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "value":"Bearer "
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateRideRoot"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Update ride path request failed"
                    },
                    "404": {
                        "description": "Update ride path request failed"
                    }
                }
            }
        },
        "/ride/end-ride": {
            "post": {
                "tags": [
                    "Ride"
                ],
                "summary": "This api trigger when ride ends by driver, and mark the ride as completed with storing the lat and long values",
                "description": null,
                "operationId": "endRide",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "value":"Bearer "
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endRide"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "End ride request failed"
                    },
                    "404": {
                        "description": "End ride request failed"
                    }
                }
            }
        }
     },
    "definitions": {
...
            "acceptRideRequest":{
                "type": "object",
                "required": [
                    "ride_id"                    
                ],
                "properties": {
                    "ride_id": {
                        "type": "string"
                    }
                }
            },
...
            "rideStartedByDriver":{
                "type": "object",
                "required": [
                    "ride_id",
                    "start_lat",
                    "start_long"
                ],
                "properties": {
                    "ride_id": {
                        "type": "string"
                    },
                    "start_lat": {
                        "type": "string"
                    },
                    "start_long": {
                        "type": "string"
                    }
                } 
            },
            "updateRideRoot":{
                "type": "object",
                "required": [
                    "ride_id",
                    "current_lat",
                    "current_long"
                ],
                "properties": {
                    "ride_id": {
                        "type": "string"
                    },
                    "current_lat": {
                        "type": "string"
                    },
                    "current_long": {
                        "type": "string"
                    }
                } 
            },
            "endRide": {
                "type": "object",
                "required": [
                    "ride_id",
                    "end_lat",
                    "end_long"
                ],
                "properties": {
                    "ride_id": {
                        "type": "string"
                    },
                    "end_lat": {
                        "type": "string"
                    },
                    "end_long": {
                        "type": "string"
                    }
                } 
            }
    }
}